# services/profile-service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Profile & Catalogue Service REST Query API",
    "description": "Query catalogue and manage profiles.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/companions": {
      "get": {
        "tags": [
          "Catalogue Query"
        ],
        "summary": "Search Companions",
        "operationId": "search_companions_api_v1_companions_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "minPrice",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minprice"
            }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Maxprice"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Pagesize"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companions/{companion_id}": {
      "get": {
        "tags": [
          "Catalogue Query"
        ],
        "summary": "Get Companion Detail",
        "operationId": "get_companion_detail_api_v1_companions__companion_id__get",
        "parameters": [
          {
            "name": "companion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Companion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/me": {
      "get": {
        "tags": [
          "Profile Management Query"
        ],
        "summary": "Get My Profile",
        "operationId": "get_my_profile_api_v1_profile_me_get",
        "parameters": [
          {
            "name": "user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Id"
            }
          },
          {
            "name": "user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Role"
            }
          },
          {
            "name": "user-status",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Status"
            }
          },
          {
            "name": "user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Profile Management Command"
        ],
        "summary": "Create My Profile",
        "operationId": "create_my_profile_api_v1_profile_me_post",
        "parameters": [
          {
            "name": "user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Id"
            }
          },
          {
            "name": "user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Role"
            }
          },
          {
            "name": "user-status",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Status"
            }
          },
          {
            "name": "user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileRequestBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profile Management Command"
        ],
        "summary": "Update My Profile",
        "operationId": "update_my_profile_api_v1_profile_me_put",
        "parameters": [
          {
            "name": "user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Id"
            }
          },
          {
            "name": "user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Role"
            }
          },
          {
            "name": "user-status",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Status"
            }
          },
          {
            "name": "user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/me/media/presigned-urls": {
      "post": {
        "tags": [
          "Media Management"
        ],
        "summary": "Request Presigned Url",
        "operationId": "request_presigned_url_api_v1_profile_me_media_presigned_urls_post",
        "parameters": [
          {
            "name": "user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Id"
            }
          },
          {
            "name": "user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Role"
            }
          },
          {
            "name": "user-status",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Status"
            }
          },
          {
            "name": "user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUrlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/me/scenarios": {
      "post": {
        "tags": [
          "Scenario Management"
        ],
        "summary": "Create Scenario",
        "operationId": "create_scenario_api_v1_profile_me_scenarios_post",
        "parameters": [
          {
            "name": "user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Id"
            }
          },
          {
            "name": "user-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Role"
            }
          },
          {
            "name": "user-status",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Status"
            }
          },
          {
            "name": "user-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User-Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScenarioRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateScenarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProfileRequestBody": {
        "properties": {
          "displayName": {
            "type": "string",
            "title": "Displayname",
            "description": "Display name for profile",
            "example": "Kano Chizuru"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio",
            "description": "Biography/Introduction text",
            "example": "Perfect rental girlfriend"
          },
          "availableCities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Availablecities",
            "description": "List of cities where active",
            "example": [
              "Hanoi",
              "HCM"
            ]
          }
        },
        "type": "object",
        "required": [
          "displayName",
          "availableCities"
        ],
        "title": "CreateProfileRequestBody"
      },
      "CreateProfileResponse": {
        "properties": {
          "companionId": {
            "type": "string",
            "title": "Companionid"
          }
        },
        "type": "object",
        "required": [
          "companionId"
        ],
        "title": "CreateProfileResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PresignedUrlRequest": {
        "properties": {
          "assetType": {
            "type": "string",
            "title": "Assettype",
            "description": "IMAGE or VOICE",
            "example": "VOICE"
          },
          "sizeBytes": {
            "type": "integer",
            "title": "Sizebytes",
            "description": "File size in bytes",
            "example": 1048576
          },
          "durationSeconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Durationseconds",
            "description": "Required for VOICE type",
            "example": 25
          },
          "contentType": {
            "type": "string",
            "title": "Contenttype",
            "description": "MIME content type",
            "example": "audio/mp3"
          }
        },
        "type": "object",
        "required": [
          "assetType",
          "sizeBytes",
          "contentType"
        ],
        "title": "PresignedUrlRequest"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "SuccessResponse"
      },
      "UpdateProfileRequestBody": {
        "properties": {
          "displayName": {
            "type": "string",
            "title": "Displayname",
            "description": "Display name for profile",
            "example": "Kano Chizuru"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio",
            "description": "Biography/Introduction text",
            "example": "Perfect rental girlfriend"
          },
          "availableCities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Availablecities",
            "description": "List of cities where active",
            "example": [
              "Hanoi",
              "HCM"
            ]
          },
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatarurl",
            "description": "Optional avatar URL",
            "example": "https://s3.rentgf.com/companion-avatar.jpg"
          }
        },
        "type": "object",
        "required": [
          "displayName",
          "availableCities"
        ],
        "title": "UpdateProfileRequestBody"
      },
...
    }
  }
}